% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_centralities.R
\name{calculate_centralities}
\alias{calculate_centralities}
\title{Calculate the centralities of a GGM}
\usage{
calculate_centralities(ega)
}
\arguments{
\item{ega}{Output of construct_ggm}
}
\value{
A dataframe containing centrality measurements for each node
}
\description{
This function takes in the output of construct_ggm and
returns a dataframe containing the betweenness, closeness,
strength, and expected influence for each node in the network.
}
