% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootEGA_analysis.R
\name{calculate_dimStab}
\alias{calculate_dimStab}
\title{Calculate item and dimension stability}
\usage{
calculate_dimStab(boot.ega)
}
\arguments{
\item{boot.ega}{Output of run_bootEGA function}
}
\value{
Output of EGAnet dimensionStability function
}
\description{
This function calculates item and dimension stability
using the EGAnet function dimensionStability
}
