% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_uva.R
\name{calculate_wTO}
\alias{calculate_wTO}
\title{Calculate wTO}
\usage{
calculate_wTO(df)
}
\arguments{
\item{df}{Dataframe representing symptom severity/frequency}
}
\value{
The list of the outputs from the UVA function as implemented
in the R package EGAnet
}
\description{
This function calculates the wTO of symptoms in a dataframe based on the
Unique Variable Analysis (UVA) method described in Christensen, Garrido,
& Colino (Multivariate Behavioral Research, 2023)
}
