% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootnet_analysis.R
\name{casedrop_boot}
\alias{casedrop_boot}
\title{Casedrop bootstrapping}
\usage{
casedrop_boot(
  df,
  normal = FALSE,
  nBoots = 2500,
  statistics = c("edge", "strength", "closeness", "betweenness"),
  communities = NULL
)
}
\arguments{
\item{df}{Dataframe of symptom severity/frequency data}

\item{normal}{Whether to consider all variables normally distrubted.
If false, conducts a non-paranormal transformation (Default: FALSE)}

\item{nBoots}{Number of bootstrapping iterations to perform (Default: 2500)}

\item{statistics}{Vector indicating which statistics to store. See bootnet
for options. (Default: c('edge','strength','closeness','betweenness')}

\item{communities}{If you are running bootnet on bridge centrality measures, use
this parameter to set community labels (set this to ega$wc, where ega is the
output from the construct_ggm function)}
}
\value{
The output of the bootnet function when run with casedrop
bootstrapping
}
\description{
This function performs casedrop bootstrapping to analyze edge weight
and centrality measure stability
}
