% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_format.R
\name{check_data_format}
\alias{check_data_format}
\title{Check if a dataframe is formatted according to PRONA specifications}
\usage{
check_data_format(df)
}
\arguments{
\item{df}{Datframe representing symptom severity/frequency}
}
\description{
This function checks if a dataframe is formatted properly for PRONA
and throws an error if not
}
\details{
Requirements include:
\enumerate{
\item First column must be named 'ID'
\item All remaining columns must be numerical values
\item There must not be any NA values in the dataframe
}
}
