% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_normality.R
\name{check_normality}
\alias{check_normality}
\title{Check normality of every variable in a symptom severity/frequency dataframe}
\usage{
check_normality(df)
}
\arguments{
\item{df}{Dataframe representing symptom severity/frequency (PRONA format)}
}
\description{
This function runs every symptom through the Shapiro-Wilk normality tests
and prints the results. A p-value < 0.05 in the Shapiro-Wilk normality test
suggests that the variable is not normally distributed. This function
is helpful in knowing whether or not to use a non-paranormal transformation
when constructing a GGM.
}
