% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised_clustering.R
\name{cluster_subsamples}
\alias{cluster_subsamples}
\title{Subsample patients and re-run clustering}
\usage{
cluster_subsamples(
  data,
  detectAlgo = "WT",
  simil_measure = "ARI",
  simplify_graphs = TRUE,
  sampling_rates = c(1, 0.99, 0.95, 0.9, 0.8, 0.7, 0.6, 0.5, 0.4, 0.3, 0.2, 0.1)
)
}
\arguments{
\item{data}{Dataframe of symptom severities (formatted as required by PRONA)}

\item{detectAlgo}{The community detection algorithm to use. Options include
FG (fast greedy), IM (infomap), LP (label propagation), LE (leading eigen),
LV (louvain), or WT (walktrap). (Default: WT)}

\item{simil_measure}{The similarity measure to use, either 'Euclidean'
or 'ARI' (Adjusted Rand Index) (Default: ARI)}

\item{simplify_graphs}{Boolean that indicates whether to simplify the graph
by removing multi-edges and loops (Default: TRUE)}

\item{sampling_rates}{Vector containing the percentages to subset the dataset
by. Default is c(1, 0.99, 0.95, 0.9, 0.8, 0.7, 0.6, 0.5, 0.4, 0.3, 0.2, 0.1).
If you change this, the vector must start with 1 (meaning the first subsample
is 100\% of the dataset) for downstream analysis to work.}
}
\value{
A dataframe containing the community membership of each patient in
each of the subsampling runs.
}
\description{
This function takes subsamples of the patient dataset and re-runs concordance
network clustering to assess the stability of patient clusters. Specifically,
it takes random subsamples of 100\%, 99\%, 95\%, 90\%, 80\%, 70\%, 60\%, 50\%, 40\%,
30\%, 20\%, and 10\% of the patient dataset. It re-runs clustering and outputs
a dataframe containing the community membership of each patient in each of
the subsampling runs.
}
