% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_visualization.R
\name{community_line_plot}
\alias{community_line_plot}
\title{Plot a line plot of the severities for the communities you specify}
\usage{
community_line_plot(data, communities = c(0))
}
\arguments{
\item{data}{Output of get_communities}

\item{communities}{Vector specifying which communities to plot. If you
include 0 in the vector, it will plot the severity for all patients
combined. (Default: c(0))}
}
\value{
A line plot of symptom severities in each community
}
\description{
This function plots a line plot of avergae severity (with 95\% CIs) for the
communities you specify
}
