% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised_clustering.R
\name{compute_cluster_stabilities}
\alias{compute_cluster_stabilities}
\title{Compute stabilities of multiple clusters across multiple subsampling runs}
\usage{
compute_cluster_stabilities(subsample_communities_df)
}
\arguments{
\item{subsample_communities_df}{The output of the cluster_subsamples function}
}
\value{
A dataframe containing the stability of each cluster in each
subsampling iteration.
}
\description{
This function computes the stabilities of the clusters in the original
clustering run across multiple subsampling runs. "Stability" is how often
data points that belong to the same cluster in the original data are still
clustered together in the subsamples (this is also known as Prediction
Strength). It is calculated by taking all the patients that belong to
one of the individual clusters, finding which communities those patients
have been grouped into in the new clustering iteration, and dividing the size
of the largest community in the new iteration by the size of original
community. For example, say clustering on a dataset of 1000 patients yields
three communities, one of 500 patients (community 1), one of 400 (community
2), and one of 100 (community 3). We rerun clustering on a subsample of the
data, and out of the 500 patients that originally belonged to community 1,
400 of them still belong to community 1,  50 belong to community 2, and 50
were randomly removed during subsampling. The "stability" of the
original community 1 for this subsampling iteration is 400/500 = 0.8
(aka 80\%).
}
