% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised_clustering.R
\name{compute_cluster_stability}
\alias{compute_cluster_stability}
\title{Compute stability of a single cluster across multiple subsampling runs}
\usage{
compute_cluster_stability(cluster_IDs, cluster_df, cluster_cols)
}
\arguments{
\item{cluster_IDs}{The IDs of the patients in the original cluster of
interest}

\item{cluster_df}{A dataframe of the cluster assignments across different
subsampling iterations for only the patients in the original cluster of
interest}

\item{cluster_cols}{The names of the columns for each clustering
iteration}
}
\value{
A vector of the stability of the cluster of iterest across each
of the subsampling iterations
}
\description{
This is a helper function for compute_cluster_stabilities. It computes the
stability of a single cluster across different runs (iterations) of
concordance network-based clustering.
}
