% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_visualization.R,
%   R/unsupervised_clustering.R
\name{create_community_heatmap}
\alias{create_community_heatmap}
\title{Function for plotting heatmap for a single community}
\usage{
create_community_heatmap(
  community_symptom_data_list,
  data,
  cluster_number,
  col_fun,
  cluster_rows,
  symptom_order
)

create_community_heatmap(
  community_symptom_data_list,
  data,
  cluster_number,
  col_fun,
  cluster_rows,
  symptom_order
)
}
\arguments{
\item{community_symptom_data_list}{list of symptom data for each community}

\item{data}{symptom severity for single community}

\item{cluster_number}{number of community of interest}

\item{col_fun}{palette to use for heatmap}

\item{cluster_rows}{Boolean representing whether or not to cluster
the rows via heirarchical clustering}

\item{symptom_order}{A vector of symptom names indicating the order
to plot symptoms on the y-axis.}
}
\value{
A heatmap of the symptom severities in one community

A heatmap of the symptom severities in one community
}
\description{
Function for plotting a heatmap for a single community.
Helper function for plot_community_heatmap.

Function for plotting a heatmap for a single community.
Helper function for plot_community_heatmap.
}
