% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_ptsd_data.R
\name{format_ptsd_data}
\alias{format_ptsd_data}
\title{Format PTSD data}
\usage{
format_ptsd_data(FullData)
}
\arguments{
\item{FullData}{Datframe representing qs.csv}
}
\value{
A dataframe of the qs.csv data reformatted to match the PRONA format
}
\description{
This function takes the qs.csv file available at
https://datashare.nida.nih.gov/study/nida-ctn-0015
and converts it to the format necessary for PRONA.
This function is provided as a reference for those
interested in directly downloading and working with
the raw data; however, running this function is not
necessary as the reformatted data is already provided
on the GitHub and within the installed package.
}
\details{
These steps are adapted from the supplmentary materials
of Epskamp, Borsboom, & Fried (2018)
DOI: 10.3758/s13428-017-0862-1

To download the raw data, go to:
https://datashare.nida.nih.gov/study/nida-ctn-0015
and click on 'CTN-0015 Data Files'. Fill out the
required data sharing agreement and navigate to the
file titled 'qs.csv'.
}
