% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised_clustering.R
\name{get_communities}
\alias{get_communities}
\title{Detect patient communities based on similarity of symptom patterns}
\usage{
get_communities(
  data,
  nrows = nrow(data),
  ncols = ncol(data),
  detectAlgo = "WT",
  simil_measure = "ARI",
  simplify_graphs = TRUE
)
}
\arguments{
\item{data}{Dataframe of symptom severities (formatted as required by PRONA)}

\item{nrows}{Number of rows in the dataframe (Defaults to the number of rows
in data)}

\item{ncols}{Number of columns in the dataframe (Defaults to the number of
columns in data)}

\item{detectAlgo}{The community detection algorithm to use. Options include
FG (fast greedy), IM (infomap), LP (label propagation), LE (leading eigen),
LV (louvain), or WT (walktrap). (Default: WT)}

\item{simil_measure}{The similarity measure to use, either 'Euclidean'
or 'ARI' (Adjusted Rand Index) (Default: ARI)}

\item{simplify_graphs}{Boolean that indicates whether to simplify the graph
by removing multi-edges and loops (Default: TRUE)}
}
\value{
A new dataframe that includes community grouping information along
with original symptom severity data
}
\description{
This function takes in PRO symptom data and detects communities
of patients with shared symptom patterns based on concordance network
clustering. See Henry et al., Plos One, 2018 for more details on
methodology.
}
