% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_visualization.R
\name{get_community_summary}
\alias{get_community_summary}
\title{Get a summary of each symptom in each community}
\usage{
get_community_summary(data)
}
\arguments{
\item{data}{output of get_communities}
}
\value{
dataframe with summary statistics for each symptom
in each community
}
\description{
Function that takes the output of get_communities and
returns a dataframe with the mean, median, upper bound,
and lower bound of each symptom in each community
}
