% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_ggm.R
\name{get_ggm_weights}
\alias{get_ggm_weights}
\title{Returns a dataframe of the weights of each edge in a GGM}
\usage{
get_ggm_weights(ega)
}
\arguments{
\item{ega}{Output from construct_ggm}
}
\value{
A dataframe of the weights of each edge in the GGM
}
\description{
This function takes the output from construct_ggm and returns
a dataframe of the weights on each edge in the GGM
}
