% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_nct.R
\name{get_NCT_pvalues}
\alias{get_NCT_pvalues}
\title{Gets p-values from running NCT}
\usage{
get_NCT_pvalues(res, what)
}
\arguments{
\item{res}{The output of the run_NCT function}

\item{what}{Defines what measure to get p-values for.
Options: "edges" or "centralities"}
}
\value{
Dataframe of p-values for either individual edges
or centrality measures
}
\description{
This function gets the p-values from running NCT
}
