% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_centralities.R
\name{make_centrality_graph}
\alias{make_centrality_graph}
\title{Plot centrality graph given a dataframe of centrality measurements}
\usage{
make_centrality_graph(centralities, bridge)
}
\arguments{
\item{centralities}{A dataframe of centrality measurements}

\item{bridge}{Boolean. Whether to graph normal centrality (FALSE)
or bridge centrality (TRUE)}
}
\value{
A plot of the centrality measurements
}
\description{
This is an internal function that is used by plot_centralities and
plot_bridge centralities to make a plot of centrality measurements
}
