% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootnet_analysis.R
\name{nodedrop_boot}
\alias{nodedrop_boot}
\title{Nodedrop bootstrapping}
\usage{
nodedrop_boot(
  df,
  normal = FALSE,
  nBoots = 2500,
  statistics = c("edge", "strength", "closeness", "betweenness"),
  gamma = 0,
  ...
)
}
\arguments{
\item{df}{Dataframe of symptom severity/frequency data}

\item{normal}{Whether to consider all variables normally distrubted.
If false, conducts a non-paranormal transformation (Default: FALSE)}

\item{nBoots}{Number of bootstrapping iterations to perform (Default: 2500)}

\item{statistics}{Vector indicating which statistics to store. See bootnet
for options. (Default: c('edge','strength','closeness','betweenness')}

\item{gamma}{EBIC tuning parameter to use (must be 0, 0.25, or 0.5)
(Default: 0)}
}
\value{
The output of the bootnet function when run with nodedrop
bootstrapping
}
\description{
This function performs nodedrop bootstrapping to analyze edge weight
and centrality measure stability. Any unspecified parameters are passed
into the bootnet function from the bootnet package.
}
