% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_visualization.R
\name{order_cols_by_mean}
\alias{order_cols_by_mean}
\title{Order columns of a dataframe from highest to lowest mean value}
\usage{
order_cols_by_mean(df)
}
\arguments{
\item{data}{symptom severity data (without ID and community columns)}
}
\value{
List of column names, ordered by mean severity
}
\description{
This function aims to order the columns of a symptom severity dataframe
from highest to lowest mean severity. It is a helper function for
community_line_plot
}
