% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_visualization.R
\name{order_rows_by_vector}
\alias{order_rows_by_vector}
\title{Function for ordering rows of a dataframe}
\usage{
order_rows_by_vector(data, order_vector)
}
\arguments{
\item{data}{dataframe of symptom severity}

\item{order_vector}{vector of how you want symptoms ordered}
}
\value{
dataframe with rows ordered as specified
}
\description{
Function for order rows of a dataframe based on an input vector.
This is a helper function create_community_heatmap
}
