% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_uva.R
\name{perform_uva}
\alias{perform_uva}
\title{Perform Unique Variable Analysis (UVA)}
\usage{
perform_uva(df, reduce = FALSE, scale = 1, output_dir = "")
}
\arguments{
\item{df}{Dataframe representing symptom severity/frequency}

\item{reduce}{Boolean. Should redundancy reduction be performed?
(default: FALSE)}

\item{scale}{The scale on which symptom severity/frequency is measured.
For example if symptoms are measured on a 0-10 scale, scale = 10.
(default: 1) (only applicable if reduce = TRUE)}

\item{output_dir}{Directory in which the .RDS and .csv outputs of UVA should
be saved (only applicable if reduce = TRUE) (defaults to current directory)}
}
\value{
The list of the outputs from the UVA function as implemented
in the R package EGAnet
}
\description{
This function performs Unique Variable Analysis (UVA) on a
dataframe of symptom severity/frequency data (in PRONA format).
The function allows the user to manually choose which variables
to consolidate
}
\details{
It saves the output from UVA in a .RDS file, and the new dataframe
with consolidated latent variables in a .csv file

For more information on UVA, see Christensen, Garrido, & Colino
(PsyArXiv, doi: 10.31234/osf.io/4kra2)
}
