% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_uva.R
\name{perform_uva_consolidation}
\alias{perform_uva_consolidation}
\title{Perform Unique Variable Analysis (UVA) consolidation}
\usage{
perform_uva_consolidation(
  df,
  scale = 1,
  cut.off = 0.25,
  reduce.method = "latent",
  new.names = NULL,
  output_dir = ""
)
}
\arguments{
\item{df}{Dataframe representing symptom severity/frequency}

\item{scale}{The scale on which symptom severity/frequency is measured.
For example if symptoms are measured on a 0-10 scale, scale = 10.
(default: 1)}

\item{cut.off}{Cut-off used to determine when pairwise wto values are
considered locally dependent (or redundant). Must be values between 0 and 1.
Defaults to 0.25.}

\item{reduce.method}{Method to reduce redundancies. Available options:
"latent", "mean", "remove", "sum". See Christensen, Garrido, & Colino
for more details (Default: latent)}

\item{new.names}{Vector of new names to give to consolidated variables.
Variable pairs will be renamed in descending order of wTO. If this vector
is not given, new variable pairs will be renamed "CV1", "CV2", etc...
Moreover, if reduce.method = "remove", this vector will not be used.
(Default: NULL)}

\item{output_dir}{Directory in which the .RDS and .csv outputs of UVA should
be saved (defaults to current directory)}
}
\value{
The list of the outputs from the UVA function as implemented
in the R package EGAnet
}
\description{
This function performs Unique Variable Analysis (UVA) consolidation
on a symptom severity dataframe, given a certain wTO cutoff.
}
\details{
For more information on UVA, see Christensen, Garrido, & Colino
(Multivariate Behavioral Research, 2023)
}
