% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootEGA_analysis.R
\name{plot_bootEGA}
\alias{plot_bootEGA}
\title{Plot median network structure from bootEGA}
\usage{
plot_bootEGA(boot.ega)
}
\arguments{
\item{boot.ega}{Output of run_bootEGA function}
}
\value{
A plot of the median GGM
}
\description{
This function plots the median determined by BootEGA stability analysis
}
