% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_centralities.R
\name{plot_bridge_centralities}
\alias{plot_bridge_centralities}
\title{Plot the bridge centralities of a GGM}
\usage{
plot_bridge_centralities(ega)
}
\arguments{
\item{ega}{Output of construct_ggm}
}
\value{
Plot of bridge strength, closeness, and betweenness centrality
}
\description{
This function takes in the output of construct_ggm and
plots the bridge strength, bridge closeness, and bridge
betweenness centrality of the network.
}
