% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootnet_analysis.R
\name{plot_casedrop_boot}
\alias{plot_casedrop_boot}
\title{Plot casedrop bootstrapping results}
\usage{
plot_casedrop_boot(casedrop.boot, type = "centralities")
}
\arguments{
\item{casedrop.boot}{Output of casedrop_boot function}

\item{type}{String indicating which measurements to plot. Options are
"edge" (which plots only casedrop results for edge weight), "centralities"
(which plots the casedrop results for strength, closeness, and betweenness),
and "bridge centralities" (which plots the casedrop results for bridge
strength, bridge closeness, and bridge betweenness) (Default: "centralities")}
}
\value{
A graph of edge weight/centrality measure stability
}
\description{
This function performs plots the results of casedrop bootstrapping
to analyze edge weight and centrality measure stability
}
