% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_centralities.R
\name{plot_centralities}
\alias{plot_centralities}
\title{Plot the centralities of a GGM}
\usage{
plot_centralities(ega)
}
\arguments{
\item{ega}{Output of construct_ggm}
}
\value{
Plot of strength, closeness, and betweenness centrality
}
\description{
This function takes in the output of construct_ggm and
plots the strength, closeness, and betweenness centrality
of the network.
}
