% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised_clustering.R
\name{plot_cluster_stabilities}
\alias{plot_cluster_stabilities}
\title{Plot cluster stabilities}
\usage{
plot_cluster_stabilities(stabilities_df)
}
\arguments{
\item{stabilities_df}{The output of compute_cluster_stabilities}
}
\value{
A ggplot object of a plot of the cluster stabilities
}
\description{
This function plots the results of compute_cluster_stabilities.
}
