% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_normality.R
\name{plot_density}
\alias{plot_density}
\title{Plot density distribution of every variable in a symptom severity/frequency dataframe}
\usage{
plot_density(df, color = "#4caf4a", x_label = "severity\\ndistribution")
}
\arguments{
\item{df}{Dataframe representing symptom severity/frequency (PRONA format)}

\item{color}{Color of the ridge plot (Default: Green)}

\item{x_label}{X-axis label (Default: "severity distribution")}
}
\value{
A ggplot2 object representing the ridge plot of symptom severity/frequency
}
\description{
This function plots a ridge plot for every variable the dataframe.
}
