% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_ggm.R
\name{plot_ggm}
\alias{plot_ggm}
\title{Plot a GGM}
\usage{
plot_ggm(ega)
}
\arguments{
\item{ega}{Output from construct_ggm}
}
\value{
A plot of the GGM
}
\description{
This function takes the output from construct_ggm and plots it
}
