% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootEGA_analysis.R
\name{plot_item_stability}
\alias{plot_item_stability}
\title{Plot item stability}
\usage{
plot_item_stability(dimStab, colors = NULL)
}
\arguments{
\item{dimStab}{Output of calculate_dimStab function}

\item{colors}{A list of colors to color the dimension stability
plot with (optional)}
}
\value{
A graph of item stability values
}
\description{
This function plots item stability using the output of the
calculate_dimStab function
}
