% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_nct.R
\name{plot_NCT}
\alias{plot_NCT}
\title{Plot Network Comparison Test (NCT)}
\usage{
plot_NCT(res, what)
}
\arguments{
\item{res}{The output of the run_NCT function}

\item{what}{Defines what has to be plotted: results pertaining
to test on invariance of global strength ("strength"), network
structure ("network"), edge strength ("edge"), or specific
centrality measure ("centrality")}
}
\value{
Plot of NCT results
}
\description{
This function plots the results of the run_NCT function
}
