% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootnet_analysis.R
\name{plot_nodedrop_boot}
\alias{plot_nodedrop_boot}
\title{Plot nodedrop bootstrapping results}
\usage{
plot_nodedrop_boot(nodedrop.boot, type = "centralities")
}
\arguments{
\item{nodedrop.boot}{Output of casedrop_boot function}

\item{type}{String indicating which measurements to plot. Options are
"edge" (which plots only casedrop results for edge weight) or "centralities"
(which plots the casedrop results for strength, closeness, and betweenness)}
}
\value{
A graph of edge weight/centrality measure stability
}
\description{
This function performs plots the results of nodedrop bootstrapping
to analyze edge weight and centrality measurement stability
}
