% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootnet_analysis.R
\name{plot_nonparam_boot}
\alias{plot_nonparam_boot}
\title{Plot nonparametric bootstrapping}
\usage{
plot_nonparam_boot(nonparam.boot)
}
\arguments{
\item{nonparam.boot}{Output of nonparam_boot function}
}
\value{
A graph of edge weight accuracy
}
\description{
This function performs plots the results of nonparametric bootstrapping
to analyze edge weight accuracy
}
