% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_severity_frequency_occurrence.R
\name{plot_occurrence}
\alias{plot_occurrence}
\title{Plot bar chart of occurrence}
\usage{
plot_occurrence(df, color = "royalblue", y_label = "occurrence", cutoff = 1)
}
\arguments{
\item{df}{Dataframe of symptom severities or frequencies (PRONA format)}

\item{color}{Color of the barplot (default blue)}

\item{y_label}{Y-axis label}

\item{cutoff}{Severity/Frequency cutoff to define occurrence at (default: 1)}
}
\value{
A ggplot2 object of the occurrence barplot
}
\description{
This function takes in a dataframe of symptom severities or frequencies
(formatted as required by PRONA) and returns a bar chart of the percent
of times each symptom occurs across the entire cohort.
}
