% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_severity_frequency_occurrence.R
\name{plot_severity}
\alias{plot_severity}
\title{Plot bar chart of severity}
\usage{
plot_severity(df, color = "#4caf4a", y_label = "severity")
}
\arguments{
\item{df}{Dataframe of symptom severities (formatted as required by PRONA)}

\item{color}{Color of the barplot (default green)}

\item{y_label}{Y-axis label}
}
\value{
A ggplot2 object of the severity barplot
}
\description{
This function takes in a dataframe of symptom
severities (formatted as required by PRONA)
and returns a bar chart of the severity for each
symptom.
}
