% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_uva.R
\name{plot_wTO}
\alias{plot_wTO}
\title{Plot weighted topological overlap (wTO) of Unique Variable Analysis (UVA)}
\usage{
plot_wTO(uva_output, color = "#ff7f00")
}
\arguments{
\item{uva_output}{List output by perform_uva}

\item{color}{Color of the barplot (Default: orange)}
}
\value{
A bar chart of the wTOs
}
\description{
Plot a bar chart of the weighted topological overlaps (wTOs) of the
perform_uva output
}
