% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootEGA_analysis.R
\name{run_bootEGA}
\alias{run_bootEGA}
\title{Perform BootEGA}
\usage{
run_bootEGA(df, normal = FALSE, iter = 10000)
}
\arguments{
\item{df}{Dataframe of symptom severity/frequency data}

\item{normal}{Whether to consider all variables normally distrubted.
If false, conducts a non-paranormal transformation (Default: FALSE)}

\item{iter}{Number of permutations for bootstrap analysis (Default: 10,000)}
}
\value{
The outboot of the BootEGA function from EGAnet
}
\description{
This function analyzes symptom cluster stability using the boot.ega function
from EGAnet
}
