% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_nct.R
\name{run_NCT}
\alias{run_NCT}
\title{Run Network Comparison Test (NCT)}
\usage{
run_NCT(
  df1,
  df2,
  normal = FALSE,
  it = 2500,
  p.adjust.methods = "none",
  test.edges = TRUE,
  test.centrality = TRUE,
  centrality = c("closeness", "betweenness", "strength", "expectedInfluence")
)
}
\arguments{
\item{df1}{First dataframe of symptom severity/frequency data}

\item{df2}{Second dataframe of symptom severity/frequency data}

\item{normal}{Boolean. Whether to consider all variables normally distrubted.
If false, conducts a non-paranormal transformation (Default: FALSE)}

\item{it}{Number of bootstrapping iterations to run (Default: 2500)}

\item{p.adjust.methods}{Character. Can be one of "holm", "hochberg",
"hommel", "bonferroni", "BH", "BY", "fdr", or "none". To control
(or not) for testing of multiple edges. Defaults to "none".}

\item{test.edges}{Boolean. Whether to test differences in individual
edge weights. (Default: TRUE)}

\item{test.centrality}{Boolean. Whether to test differences in centrality
measures (Default: TRUE)}

\item{centrality}{Vector of which centrality measures to test (Default:
c('closeness','betweenness','strength','expectedInfluence'))}
}
\value{
The output of the NCT function from the NetworkComparisonTest
package
}
\description{
This function runs a network comparison test between two dataframes
of symptom severity data
}
