% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised_clustering.R
\name{similarity_matrix}
\alias{similarity_matrix}
\title{Compute a similarity matrix based on an input matrix}
\usage{
similarity_matrix(mat, concords = vecs2concords(mat), simil_measure = "ARI")
}
\arguments{
\item{mat}{The input matrix for which the similarity matrix
will be calculated}

\item{concords}{By default, this argument takes the value returned by
the vec2concords(mat) function, which calculates the concordance
matrices for each row of mat}

\item{simil_measure}{The similarity measure to use, either 'Euclidean'
or 'ARI' (Adjusted Rand Index) (Default: ARI)}
}
\description{
This function computes a similarity matrix based on an input
matrix (mat) using either Euclidean disease or the Adjusted
Rand Index (ARI) as the similarity measure. See Henry et al.,
Plos One, 2018 for more details. This is a helper function
for the get_communities function.
}
