% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootnet_analysis.R
\name{summarize_nonparam_boot}
\alias{summarize_nonparam_boot}
\title{Get nonparam boot summary table}
\usage{
summarize_nonparam_boot(nonparam.boot)
}
\arguments{
\item{nonparam.boot}{Output of nonparam_boot function}
}
\value{
A dataframe summarizing the nonparametric bootstrapping
results
}
\description{
This function returns a summary table of the object that
is outputted from the nonparam_boot function
}
