% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised_clustering.R
\name{vec2concord}
\alias{vec2concord}
\title{Calculate the concordance matrix for a vector}
\usage{
vec2concord(vec)
}
\arguments{
\item{vec}{input vector}
}
\description{
This function takes a vector (vec) and computes a concordance
matrix using the formula (vec*vec^T)/100. The result of this
function is a square matrix that represents the concordance
between elements of the vector. This is a helper function for
the vec2concords function
}
