% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised_clustering.R
\name{vecs2concords}
\alias{vecs2concords}
\title{Calculate a concordance matrix for each row in a matrix}
\usage{
vecs2concords(mat)
}
\arguments{
\item{mat}{input matrix}
}
\description{
This function takes a matrix (mat) where each row represents
a vector. It then applies the vec2concord function to each row
in the matrix, resulting in a list of concordance matrices for
each row in mat. This is a helper function for the similarity_matrix
function
}
