% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_visualization.R
\name{plot_community_heatmap}
\alias{plot_community_heatmap}
\title{Plot heatmap of symptom severities in different communities}
\usage{
plot_community_heatmap(
  data,
  cluster_rows = TRUE,
  symptom_order = NULL,
  colors = "red"
)
}
\arguments{
\item{data}{The output of the get_communities function, which is
a dataframe of symptom severities with a column representing which
community each patient belongs to}

\item{cluster_rows}{Boolean representing whether or not to cluster
the rows via heirarchical clustering (Default: TRUE)}

\item{symptom_order}{A vector of symptom names indicating the order
to plot symptoms on the y-axis. This order will only show up if
cluster_rows = FALSE. The default order if cluster_rows = FALSE is
the order of column names in the dataframe (minus the ID and
community columns)}

\item{colors}{A vector of colors to use for plotting the heatmaps
for each community. If there is only one color, all communities will
be plotted in that color. If the vector denotes colors for each community,
each community will be a different color. If the length of the vector
and the number of communities do not match, an error will be thrown.
(Default: 'red').}
}
\value{
A heatmap of the symptom severities in each community
}
\description{
This function takes in the output of the get_communities function
and plots a heatmap of the symptom severities in each community
}
