% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_visualization.R
\name{plot_community_heatmap}
\alias{plot_community_heatmap}
\title{Plot heatmap of symptom severities in different communities}
\usage{
plot_community_heatmap(
  data,
  cluster_rows = TRUE,
  symptom_order = NULL,
  network = NULL
)
}
\arguments{
\item{data}{The output of the get_communities function, which is
a dataframe of symptom severities with a column representing which
community each patient belongs to}

\item{cluster_rows}{Boolean representing whether or not to cluster
the rows via heirarchical clustering (Default: TRUE)}

\item{symptom_order}{A vector of symptom names indicating the order
to plot symptoms on the y-axis. This overrides cluster_rows.
The default order if cluster_rows = FALSE and symptom_order = NULL
is the order of column names in the dataframe (minus the ID and
community columns)}

\item{network}{An optional parameter that lets you pass in a network
object output by construct_ggm. If a network is passed into this function,
the heatmap will automatically order rows (symptoms) by the symptom
clusters identified in the network. This overrides the cluster_rows and
symptom_order parameters.}
}
\value{
A heatmap of the symptom severities in each community
}
\description{
This function takes in the output of the get_communities function
and plots a heatmap of the symptom severities in each community
}
