% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_ggm.R
\name{plot_ggm}
\alias{plot_ggm}
\title{Plot a GGM}
\usage{
plot_ggm(ega, colors = NULL, legend.names = NULL)
}
\arguments{
\item{ega}{Output from construct_ggm}

\item{colors}{A list of colors to color the clusters with (optional)}

\item{legend.names}{A list of strings to label the clusters with (optional)}
}
\value{
A plot of the GGM
}
\description{
This function takes the output from construct_ggm and plots it
}
